requirements.txt
pymysql
flask-mysql
flask_cors
